Relocate scan and firewall footers
