--- projects/plugins/protect/src/js/components/firewall-footer/stories/broken/index.stories.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import FirewallFooter from '../index.jsx';
-
-export default {
-	title: 'Plugins/Protect/Firewall Footer',
-	component: FirewallFooter,
-};
-
-export const Default = args => {
-	return <FirewallFooter { ...args } />;
-};
