--- projects/plugins/protect/src/js/components/scan-footer/styles.module.scss
@@ -1,4 +1,0 @@
-.product-section, .info-section {
-	margin-top: calc( var( --spacing-base ) * 7 ); // 56px
-	margin-bottom: calc( var( --spacing-base ) * 7 ); // 56px
-}
