+++ projects/plugins/protect/src/js/routes/firewall/styles.module.scss
@@ -190,3 +190,34 @@
         margin-bottom: 0;
     }
 }
+
+.footer-checkbox {
+    display: flex;
+    align-items: center;
+
+    & input[type=checkbox]:checked {
+        background: var( --jp-green-40 );
+        border-color: var( --jp-green-40 );
+    }
+
+    & input[type=checkbox]:indeterminate {
+        background: var( --jp-green-40 );
+        border-color: var( --jp-green-40 );
+    }
+
+    & input[type=checkbox]:focus {
+        box-shadow: 0 0 0 2px #fff, 0 0 0 4px var( --jp-green-40 );
+    } 
+}
+
+.standalone-mode-section, .share-data-section {
+	margin-top: calc( var( --spacing-base ) * 7 ); // 56px
+    margin-bottom: calc( var( --spacing-base ) * 7 ); // 56px
+}
+
+@media ( max-width: 599px ) {
+
+    .share-data-section {
+        margin-top: 0;
+    }
+}
