+++ projects/plugins/protect/src/js/routes/scan/styles.module.scss
@@ -12,3 +12,8 @@
 		width: 100%;
 	}
 }
+
+.product-section, .info-section {
+	margin-top: calc( var( --spacing-base ) * 7 ); // 56px
+	margin-bottom: calc( var( --spacing-base ) * 7 ); // 56px
+}
